@@ -11,20 +11,19 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.18</version>
+		<version>3.3.2</version>
 		<relativePath />
 	</parent>
 
 	<properties>
-		<maven.compiler.target>1.8</maven.compiler.target>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<spring.security.version>5.1.10.RELEASE</spring.security.version>
+		<maven.compiler.target>17</maven.compiler.target>
+		<maven.compiler.source>17</maven.compiler.source>
+		<!--<spring.security.version>5.1.10.RELEASE</spring.security.version>-->
 		<ojdbc.version>19.3.0.0</ojdbc.version>
-		<hibernate.version>5.4.3.Final</hibernate.version>
-		<jetty.version>9.4.44.v20210927</jetty.version>
-		<kafka.version>2.0.0</kafka.version>
+		<hibernate.version>7.0.0.Final</hibernate.version>
+		<!--<jetty.version>9.4.44.v20210927</jetty.version>-->
+		<kafka.version>2.6.3</kafka.version>
 		<protobuf.version>3.16.3</protobuf.version>
-		<swagger-version>2.0.9</swagger-version>
 		<skipITs>true</skipITs>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 	</properties>
