@@ -6,6 +6,7 @@ import java.io.IOException;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.time.Duration;
 import java.util.Properties;
 import java.util.UUID;
 
@@ -13,7 +14,7 @@ import javax.sql.DataSource;
 
 import no.nav.aura.envconfig.spring.SpringOracleUnitTestConfig;
 
-import org.apache.commons.dbcp.BasicDataSource;
+import org.apache.commons.dbcp2.BasicDataSource;
 import org.flywaydb.core.api.output.MigrateResult;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -54,10 +55,10 @@ public abstract class TestDatabaseHelper {
             ds.setUrl(url);
             ds.setUsername(username);
             ds.setPassword(password);
-            ds.setMaxWait(20000);
+            ds.setMaxWait(Duration.ofMillis(20000));
             ds.setMinIdle(0);
-            ds.setMinEvictableIdleTimeMillis(10000);
-            System.out.println("using database " + ds.getUsername() + "@" + ds.getUrl());
+            ds.setMinEvictableIdle(Duration.ofMillis(10000));
+            System.out.println("using database " + ds.getUserName() + "@" + ds.getUrl());
 
             Resource script = new DefaultResourceLoader().getResource("org/springframework/session/jdbc/schema-h2.sql");
             System.out.println("Using spring session script " + script.getFilename());
@@ -73,10 +74,10 @@ public abstract class TestDatabaseHelper {
             ds.setUrl(url);
             ds.setUsername(username);
             ds.setPassword(password);
-            ds.setMaxWait(20000);
+            ds.setMaxWait(Duration.ofMillis(20000));
             ds.setMinIdle(0);
-            ds.setMinEvictableIdleTimeMillis(10000);
-            System.out.println("using database " + ds.getUsername() + "@" + ds.getUrl());
+            ds.setMinEvictableIdle(Duration.ofMillis(10000));
+            System.out.println("using database " + ds.getUserName() + "@" + ds.getUrl());
             return ds;
         }
     }
